Main window: Fixed warning with geometry parsing.
We will no longer attempt to position the window for number of reasons.
Most important one is that GTK+ geometry handling has been deprecated
since 3.20 version and in addition to that geometry handling is window
management job.

Instead we set default size and let the rest be handled by the system
itself.