File list: Make executing files optional.
I have many executable shell and Python scripts. When I activate them
in a file manager, I always want to open them as text files, never to
execute them. I like how GNOME Files handles this: it has an option to
always execute, always open, or ask every time. This commit adds
a similar option to Sunflower, except that "ask every time" is not yet
implemented (because I don't need it) -- but it should be easy to add
later.